Move AddressTagDb to Core
