+++ wallet-test/src/test/scala/org/bitcoins/wallet/models/AddressTagDAOTest.scala
@@ -5,6 +5,7 @@ import java.sql.SQLException
 import org.bitcoins.core.wallet.utxo.StorageLocationTag.HotStorage
 import org.bitcoins.core.wallet.utxo.{
   AddressTag,
+  AddressTagDb,
   UnknownAddressTag,
   UnknownAddressTagName,
   UnknownAddressTagType
