+++ wallet/src/main/scala/org/bitcoins/wallet/api/WalletApi.scala
@@ -20,15 +20,15 @@ import org.bitcoins.core.protocol.transaction.{
 }
 import org.bitcoins.core.util.FutureUtil
 import org.bitcoins.core.wallet.fee.FeeUnit
-import org.bitcoins.core.wallet.utxo._
+import org.bitcoins.core.wallet.utxo.{
+  AddressTag,
+  AddressTagDb,
+  AddressTagType,
+  TxoState
+}
 import org.bitcoins.crypto.DoubleSha256DigestBE
 import org.bitcoins.wallet.WalletLogger
-import org.bitcoins.wallet.models.{
-  AddressDb,
-  AddressTagDb,
-  ScriptPubKeyDb,
-  TransactionDb
-}
+import org.bitcoins.wallet.models.{AddressDb, ScriptPubKeyDb, TransactionDb}
 
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.{Failure, Success}
